refactor(logic): TicketBuyer의 기능을 TicketBuilder로 이전
