+++ src/main/kotlin/lotto/domain/Ticket.kt
@@ -2,16 +2,12 @@ package lotto.domain
 
 class Ticket(val numbers: Set<Int>) {
     init {
-        require(numberIsInRange(TICKET_NUMBER_MIN, TICKET_NUMBER_MAX))
-        require(isSize(TICKET_NUMBER_SIZE))
+        require(isNumberInRange(TICKET_NUMBER_MIN, TICKET_NUMBER_MAX))
+        require(numbers.size == TICKET_NUMBER_SIZE) { "로또의 개수는 6개 입니다." }
     }
 
-    private fun numberIsInRange(min: Int, max: Int): Boolean {
+    private fun isNumberInRange(min: Int, max: Int): Boolean {
         return numbers.all { it in min until max }
-    }
-
-    private fun isSize(size: Int): Boolean {
-        return numbers.size == size
     }
 
     fun countMatches(ticket: Ticket): Int {
