--- src/main/kotlin/lotto/domain/TicketBuyer.kt
@@ -1,11 +1,0 @@
-package lotto.domain
-
-class TicketBuyer {
-    companion object {
-        fun howMuchTickets(money: Int): Int {
-            return money / TICKET_COST
-        }
-
-        const val TICKET_COST = 1000
-    }
-}
